Update sols_microscope_gui.py
Upgraded user experience -> spinbox values are now updated when mouse is 'hovered' over the acquisition frame. This results in up to date values being passed to the acquisition without the need to explicitly hit 'return' or 'focusout' of the cell. This work around avoids adding keystroke or time delay validation directly on the spinbox (which can be complex to code and/or irritating to use).