Update sols_microscope.py
The method "_save_metadata()" (which creates the 'on disc' metadata text file) now saves the 'tuple' of the microscope attributes "self.channels_per_slice" and "self.power_per_channel". This allows flexibility in the API for creating and passing other iterables (like lists) to 'apply_settings()', but ensures the metadata output is in a consistent format for downstream processing.