Update sols_microscope_gui.py
Added convenience buttons to GuiFocusPiezo to quickly center the piezo, and move up and down in 5um steps. Here 'up' and 'down' are referenced to how the image moves in the live preview window for an intuitive feel (which is opposite to the numerical value of the piezo position). The piezo slider direction was also reversed to mirror this more 'intuitive' feel.