Update sols_microscope.py
The default behavior for scipy.ndimage.zoom is to 'fill values beyond the edge' with 'cval' (=0 by default). So for certain combinations of camera chip cropping, scan range and voxel aspect ratio, an empty row (or column) of pixels would sometimes appear at the edge of the preview images. This was fine from a user perspective but was capable of upsetting the software autofocus routine (since it would confuse the 'sharp' edge from the zero pixel row as the max gradient of the fluorescent intensity!). The new argument mode='nearest' fills any empty pixels at the edges with the nearest pixel and should fix the problem.