Update sols_microscope_gui.py
Major structural update to improve logical flow, modularity and clarity. Sub GUI's that relate to a particular piece of hardware are now nested inside the 'GuiMicroscope' class which houses the more complex/interwoven 'acquisition' functions. GuiMicroscope initializes the microscope with the settings from the gui and sub gui's, taking care to avoid XYZ motion and ensure that the gui matches the state of the hardware. In addition there are many minor updates to style and format (like 'tki_cw' to 'tkcw').