Update sols_microscope.py
A change in the API -> 'filter_wheel_position' has been replaced with 'emission_filter' which takes a string argument that maps the filter name (part number) to a position in the wheel using the dictionary 'emission_filter_options'. This makes the choice of filter obvious for the user and records the filter name in the metadata. The filter specifications can then be found directly from the part numbers.
The attribute 'self.dichroic_mirror' has been added and is also saved in the metadata.