Update sols_microscope_gui.py
Shortened 'acquisition' to 'acquire' throughout. The default label is now just 'sols' (not 'sols_gui'). The 'Snap volume and save' button continues to add '_snap' to the filename and now the 'Run acquire' button adds '_acquire' to help clarify the origin of the data in the 'sols_gui_session' folder.