Update sols_microscope_gui.py
Moved the stop/cancel code in 'run_set_grid_location', 'run_move_to_grid_location' and 'run_move_to_tile' to the start of the loop to ensure the .after call is stopped immediately. This should aviod the race condition where the .after can reschedule a call to the function (and pass the 'if self.running...' guard) when it should have stopped already.