Update sols_microscope_gui.py
An important structural update was applied here (to the grid and tile navigator) which I believe avoids bugs/crashing/concurrency problems with tkinter. It seems like 'best practice' is to initialize as much as possible (before the mainloop) and then 'destroy' at the end (i.e. avoid initializing/destroying variables, frames, widgets etc as much as possible during the runtime since this can cause freezing). For example, an update here is to create 'popups' using 'tk.Toplevel' in advance, and then hide/show them with '.withdraw/.deiconify' (instead of creating and destroying). This seems faster and more stable during testing.

Otherwise mostly a tidy of the 'grid navigator':
- The grid parameters (rows, cols, XY_positions) are now consolidated into a single list (self.grid_list) to simplify the code and associated methods.
- Many internal functions/methods renamed with leading '_' (underscore) and some other renaming for improved consistency with the rest of the module.
- An unused 'move' method was removed from 'init_XY_stage'.