Update sols_microscope.py
Focus piezo change from a 100um ultrafast 'bucket' style piezo to a 400um high speed 'traditional' style piezo. The bucket style is faster but it makes it hard to adjust the correction collar on the objective (the objective needs removing each time which is slow and awkward). The extra range on the 400um piezo may also prove useful for finding samples.